+++ enum_string_extension_test/lib/basic_class.dart
@@ -16,6 +16,10 @@ class AppLocalizations {
       value1: 'Value 1 in english',
       value2: 'Value 2 in english',
       value3: 'Value 3 in english',
+      testValue1: 'TestValue 1 in english',
+      testValue2: 'TestValue 2 in english',
+      test2Value1: 'Test2Value 1 in english',
+      test2Value2: 'Test2Value 2 in english',
     )
   };
 
@@ -30,11 +34,24 @@ class AppLocalizationsLabels {
     this.value1,
     this.value2,
     this.value3,
+    this.testValue1,
+    this.testValue2,
+    this.test2Value1,
+    this.test2Value2,
   });
 
   final String value1;
   final String value2;
   final String value3;
+  final String testValue1;
+  final String testValue2;
+  final String test2Value1;
+  final String test2Value2;
+}
+
+enum TestEnum4 {
+  value1,
+  value3,
 }
 
 enum TestEnum3 {
@@ -61,10 +78,42 @@ class BasicClass {
     this.test2,
     this.test3,
     this.test4,
+    this.test5,
+    this.test6,
+    this.test7,
+    this.test8,
+    this.test9,
+    this.test10,
+    this.test11,
   });
 
+  // Should generate text() for TestEnum
   final TestEnum test;
+  // Should generate text() TestEnum2
   final TestEnum2 test2;
+  // Should generate text() for TestEnum3
   final List<TestEnum3> test3;
+
+  // Should generate testText() for TestEnum
+  @EnumKey(prefix: 'test')
   final TestEnum test4;
+// Should generate testText() for TestEnum
+  @EnumKey(prefix: 'test2')
+  final TestEnum test5;
+
+  // Should not generate code
+
+  // Duplicate prefix
+  final TestEnum test6;
+  // Should not generate duplicate text() for TestEnum
+  @EnumKey(prefix: 'test')
+  final TestEnum test7;
+  @EnumKey(prefix: 'Test')
+  final TestEnum test8;
+  // Not enum fields
+  final String test9;
+  final List<String> test10;
+  // Excluded enum
+  @EnumKey(exclude: true)
+  final TestEnum4 test11;
 }