+++ enum_string_extension_gen/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.3
+
+* Add support for `EnumKey` with prefix parameter and exclude option
+* Exclude parameter prevent code gen
+* Prefix parametr creates an changes the prefix of the text() and adds the prefix to the AppLocalization fields
+
 ## 0.0.2 Fix multiple enum and support subtypes for `List<T>`
 
 * Fix bug that prevented multiple enum to generate code